Obstacel + word file
- Added inheritance to the obstacles
- final small changes
- finished word file

+++ _FRAMEWORK/source/framework/EliteAI/EliteGraphs/EliteGraphUtilities/EGraphRenderer.cpp
@@ -19,7 +19,7 @@ namespace Elite
 				if (renderIntegrationCost)
 					integrationTxt = GetIntegrationFieldText(node);
 
-				RenderRectNode(worldPos, costTxt, integrationTxt, cellSize, GetNodeColor(node));
+				RenderRectNode(worldPos, costTxt, integrationTxt, static_cast<float>(cellSize), GetNodeColor(node));
 			}
 			else
 			{
